Time: 0 ms (100.00%) | Memory: 7.9 MB (94.86%) - LeetSync

+++ 1725-number-of-sets-of-k-non-overlapping-line-segments/number-of-sets-of-k-non-overlapping-line-segments.cpp
@@ -35,6 +35,6 @@ public:
 
         long long inverseDenominator = modPow(denominator, MOD - 2);
 
-        return static_cast<int>(numerator * inverseDenominator % MOD);
+        return (numerator * inverseDenominator % MOD);
     }
 };
